backend/requirements.txt
fastapi
uvicorn
pypdf
python-docx
pymongo
python-dotenv
python-multipart
google-genai
pydantic-settings
motor